Rename VoteSuccessfulModal to StakeSuccessfulModal
