+++ src/locales/en/common.json
@@ -684,7 +684,7 @@
   "Votes": "Votes",
   "Votes are submitted": "Votes are submitted",
   "Voting Summary": "Voting Summary",
-  "Voting confirmation": "Voting confirmation",
+  "Staking confirmation": "Staking confirmation",
   "Voting queue": "Voting queue",
   "Wallet": "Wallet",
   "Wallet address:": "Wallet address:",
