+++ src/modules/pos/validator/components/StakeStatus/Status.js
@@ -4,7 +4,7 @@ import { getTransactionStatus } from '@transaction/configuration/statusConfig';
 import { txStatusTypes } from 'src/modules/transaction/configuration/txStatus';
 import statusMessages from './statusMessages';
 import styles from './styles.css';
-import VoteSuccessfulModal from '../VoteSuccessfulModal';
+import VoteSuccessfulModal from '../StakeSuccessfulModal';
 
 const Status = ({ account, transactions, statusInfo, t, dposToken }) => {
   const status = getTransactionStatus(account, transactions, account?.summary.isMultisignature);
