+++ src/modules/pos/validator/components/StakeSuccessfulModal/StakeSuccessfulModal.css
@@ -1,10 +1,10 @@
 .successContainer {
-  display: 'flex';
+  display: flex;
   flex-direction: column;
   align-items: center;
 
   & header {
-    @mixin voteStatusHeading;
+    @mixin stakeStatusHeading;
 
     justify-content: center !important;
     padding-bottom: 4px !important;
