+++ src/modules/pos/validator/components/StakeSuccessfulModal/index.js
@@ -1,0 +1,4 @@
+import { withRouter } from 'react-router';
+import StakeSuccessfulModal from './StakeSuccessfulModal'
+
+export default withRouter(StakeSuccessfulModal);
