--- src/modules/pos/validator/components/VoteSuccessfulModal/index.js
@@ -1,4 +1,0 @@
-import { withRouter } from 'react-router';
-import VoteSuccessfulModal from './VoteSuccessfulModal'
-
-export default withRouter(VoteSuccessfulModal);
